Use UInt64 for balance constants to avoid repeatedly creating new UInt64 instances for them. (#3367)

+++ data/beaconrestapi/src/integration-test/java/tech/pegasys/teku/beaconrestapi/v1/beacon/GetStateValidatorBalancesIntegrationTest.java
@@ -56,8 +56,7 @@ public class GetStateValidatorBalancesIntegrationTest
             response.body().string(), GetStateValidatorBalancesResponse.class);
     assertThat(body.data)
         .containsExactly(
-            new ValidatorBalanceResponse(
-                UInt64.valueOf(1), UInt64.valueOf(Constants.MAX_EFFECTIVE_BALANCE)));
+            new ValidatorBalanceResponse(UInt64.valueOf(1), Constants.MAX_EFFECTIVE_BALANCE));
   }
 
   @Test
